+++ lib/page/home/home.dart
@@ -1,8 +1,9 @@
+
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'bloc/ImageListBloc.dart';
-import 'component/image_preview_view.dart';
+import 'bloc/image_list_bloc.dart';
+import 'home_body.dart';
 
 class HomeArguments {
   final String query;
@@ -10,7 +11,7 @@ class HomeArguments {
   const HomeArguments(this.query);
 }
 
-class MyHomePage extends StatefulWidget {
+class MyHomePage extends StatelessWidget {
   static const routerName = "/";
 
   final HomeArguments? args;
@@ -24,36 +25,28 @@ class MyHomePage extends StatefulWidget {
   }
 
   @override
-  State<MyHomePage> createState() => _MyHomePageState();
+  Widget build(BuildContext context) {
+    return BlocProvider(
+      create: (context) => ImageListBloc.create(),
+      child: _MyHomePageStateWidget(title, queryTarget),
+    );
+  }
 }
 
-class _MyHomePageState extends State<MyHomePage> {
-  final _imageListBloc = ImageListBloc(ImageListState.initState());
+class _MyHomePageStateWidget extends StatefulWidget {
+  final String title;
+  final String query;
+
+  const _MyHomePageStateWidget(this.title, this.query);
+
 
   @override
-  void initState() {
-    super.initState();
-
-    _startFetchImage(widget.queryTarget);
+  State<StatefulWidget> createState() {
+    return _MyHomePageBlocState();
   }
+}
 
-  _startFetchImage(String query) {
-    _imageListBloc.add(QueryEvent(query));
-  }
-
-  _nextPage() {
-    _imageListBloc.add(LoadMoreEvent());
-  }
-
-  _onClickSearch() async {
-    final result = await Navigator.pushNamed(context, '/search');
-    if (result == null) return;
-    if (!mounted) return;
-
-    if (result is String) {
-      _startFetchImage(result);
-    }
-  }
+class _MyHomePageBlocState extends State<_MyHomePageStateWidget> {
 
   @override
   Widget build(BuildContext context) {
@@ -71,28 +64,22 @@ class _MyHomePageState extends State<MyHomePage> {
               icon: const Icon(Icons.search))
         ],
       ),
-      body: BlocBuilder(
-        bloc: _imageListBloc,
-        builder: (context, state) {
-          var imageState = state as ImageListState;
-          return Stack(
-            children: [
-              Opacity(
-                opacity: imageState.loading ? 0.3 : 1.0,
-                child: IgnorePointer(
-                  ignoring: imageState.loading,
-                  child: ImagePreviewList(voList: imageState.voList, onScrollBottom: _nextPage,),
-                ),
-              ),
-              Center(
-                child: Visibility(
-                    visible: imageState.loading, child: const CircularProgressIndicator()),
-              )
-            ],
-          );
-        },
-
-      ),
+      body: const HomeBody(),
     );
+  }
+
+  _onClickSearch() async {
+    final result = await Navigator.pushNamed(context, '/search');
+    if (!mounted) return null;
+
+    if (result is String) {
+      BlocProvider.of<ImageListBloc>(context).add(QueryEvent(result));
+    }
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    BlocProvider.of<ImageListBloc>(context).add(QueryEvent(widget.query));
   }
 }